+++ docs/src/examples/components/Accordion/Performance/AccordionMinimal.perf.tsx
@@ -3,7 +3,7 @@ import * as React from 'react'
 
 const AccordionMinimalPerf = () => <Accordion />
 
-AccordionMinimalPerf.iterations = 5000
+AccordionMinimalPerf.iterations = 1000
 AccordionMinimalPerf.filename = 'AccordionMinimal.perf.tsx'
 
 export default AccordionMinimalPerf