+++ docs/src/examples/components/Alert/Performance/AlertMinimal.perf.tsx
@@ -3,7 +3,7 @@ import * as React from 'react'
 
 const AlertMinimalPerf = () => <Alert />
 
-AlertMinimalPerf.iterations = 5000
+AlertMinimalPerf.iterations = 1000
 AlertMinimalPerf.filename = 'AlertMinimal.perf.tsx'
 
 export default AlertMinimalPerf