+++ docs/src/examples/components/Button/Performance/index.tsx
@@ -8,7 +8,7 @@ const Performance = () => (
     <ComponentPerfExample
       title="Default"
       description="A default test."
-      examplePath="components/Button/Performance/ButtonSlots.perf"
+      examplePath="components/Button/Performance/ButtonMinimal.perf"
     />
   </NonPublicSection>
 )