+++ docs/src/examples/components/Header/Performance/index.tsx
@@ -11,9 +11,9 @@ const Performance = () => (
       examplePath="components/Header/Performance/HeaderSlots.perf"
     />
     <ComponentPerfExample
-      title="Default"
-      description="A test with a description shorthand."
-      examplePath="components/Header/Performance/HeaderDescription.perf"
+      title="Minimal"
+      description="Header with no props."
+      examplePath="components/Header/Performance/HeaderMinimal.perf"
     />
   </NonPublicSection>
 )