+++ docs/src/examples/components/Loader/Performance/LoaderMinimal.perf.tsx
@@ -3,7 +3,7 @@ import * as React from 'react'
 
 const LoaderMinimalPerf = () => <Loader />
 
-LoaderMinimalPerf.iterations = 5000
+LoaderMinimalPerf.iterations = 1000
 LoaderMinimalPerf.filename = 'LoaderMinimal.perf.tsx'
 
 export default LoaderMinimalPerf